Testing

+++ server/controller/todoController.js
@@ -2,12 +2,38 @@ const db = require('../models/userModel');
 
 const todoController = {};
 
+
+todoController.getTask = async (req, res, next) => {
+  const text = "SELECT * FROM public.tasks WHERE user_id=$1"
+  const { userID } = req.body;
+  try {
+    const response = await db.query(text, [userID]);
+    console.log("GET for tasks successful");
+
+    const { rows } = response;
+    rows.sort((a, b) => a._id - b._id)
+    const entries = [];
+    rows.forEach(entry => entries.push(entry.task));
+    console.log('response', entries);
+    res.locals.tasks = entries;
+    return next();
+
+  } catch (err) {
+    console.error(err)
+    return next({
+      log: "Express error handler caught an error in the getTask middleware",
+      status: 500,
+      message: { err: "An error occurred in the getTask middleware" }
+    })
+  }
+}
+
 todoController.addTask = async (req, res, next) => {
-  const { task } = req.body;
+  const { task, userID } = req.body;
   const text = `INSERT INTO public.tasks VALUES (DEFAULT, $1, $2)`
   console.log('entered middleware new task')
   try {
-    await db.query(text, [task, '1']);
+    await db.query(text, [task, userID]);
     console.log('New task created!')
     return next();
   } catch (err) {
@@ -21,11 +47,11 @@ todoController.addTask = async (req, res, next) => {
 };
 
 todoController.updateTask = async (req, res, next) => {
-  const { task, oldVal } = req.body;
-  const text = `UPDATE public.tasks SET task=$1 WHERE tasks.task = $2`
+  const { task, oldVal, userID } = req.body;
+  const text = `UPDATE public.tasks SET task=$1 WHERE tasks.task = $2 AND user_id = $3`
   console.log('entered middleware update task')
   try {
-    await db.query(text, [task, oldVal]);
+    await db.query(text, [task, oldVal, userID]);
     console.log('Task updated!')
     return next();
   } catch (err) {
@@ -39,11 +65,11 @@ todoController.updateTask = async (req, res, next) => {
 };
 
 todoController.deleteTask = async (req, res, next) => {
-  const { task } = req.body;
-  const text = `DELETE FROM public.tasks WHERE task=$1`
+  const { task, userID } = req.body;
+  const text = `DELETE FROM public.tasks WHERE task=$1 AND user_id=$2`
   console.log('entered middleware delete task')
   try {
-    await db.query(text, [task]);
+    await db.query(text, [task, userID]);
     console.log('Task deleted!')
     return next();
   } catch (err) {
@@ -56,20 +82,4 @@ todoController.deleteTask = async (req, res, next) => {
   }
 };
 
-// todoController.getTask = async (req, res, next) => {
-//   const text = "SELECT * FROM public.tasks"
-//   try {
-//     const response = await db.query(text)
-//     console.log("GET for tasks successful")
-//     res.locals.tasks = response.rows
-//     return next();
-//   } catch (err) {
-//     console.error(err)
-//     return next({
-//       log: "Express error handler caught an error in the getTask middleware",
-//       status: 500,
-//       message: { err: "An error occurred in the getTask middleware" }
-//     })
-//   }
-// }
 module.exports = todoController
